requirements.txt
pandas~=2.2

requests~=2.32
beautifulsoup4~=4.12

sqlalchemy~=2.0
psycopg2-binary~=2.9

google-auth~=2.36
google-api-python-client~=2.152

pytest~=8.0
pytest-cov~=5.0
requests-mock~=1.12

python-dotenv~=1.0